feito o modal confirmar exlusao e iniciado o modulo orcamento

+++ edit.php
@@ -44,6 +44,12 @@
     $query = "SELECT id_type, type FROM type_product";
     $sqlListTypes = mysqli_query($conn, $query);
 
+    $sqlDeleteType = "SELECT * from type_product WHERE type = '$id_type'";
+    $deleteType = mysqli_query($conn, $sqlDeleteType);
+    $rowDeleteType = mysqli_fetch_assoc($deleteType);
+
+    
+
 
 
 
@@ -73,7 +79,6 @@
 
         <div class="bg-img" style="background-image: url('./img/background.jpg');">
             <div class="overlay"></div>
-            <!-- <img src="" class="formatimg" alt="contest"> -->
 
         </div>
 
@@ -87,43 +92,41 @@
                 <form action="editestoque.php" method="POST" enctype="multipart/form-data" class="mx-md-5">
                     <input type="hidden" name="id_product" value="<?php echo $rowProduct['id_product']; ?>">
                     <div class="form-group format mx-auto inputbox">
-                        <!-- <label for="name" class="col-sm-2 col-form-label mx-auto"></label> -->
                         <div class=" ">
                             <input required type="text" class="form-control mx-auto" name="name_product" placeholder="Nome produto" required value="<?php echo $rowProduct['name_product']; ?>">
                         </div>
                     </div>
 
                     <div class="form-group format mx-auto inputbox">
-                        <!-- <label for="email" class="col-sm-2 col-form-label mx-auto"></label> -->
                         <div class="">
-                            <input required type="text" class="form-control mx-auto" name="price_product" placeholder="Preço" required value="<?php echo $rowProduct['price_product'];?>">
+                            <input required type="text" class="form-control mx-auto" name="price_product" placeholder="Preço" required value="<?php echo $rowProduct['price_product']; ?>">
                         </div>
                     </div>
 
                     <div class="form-group format mx-auto inputbox">
-                        <!-- <label for="inputEmail3" class="col-sm-2 col-form-label mx-auto"></label> -->
                         <div class="">
                             <input required type="text" class="form-control mx-auto" name="quantity" placeholder="Quantidade" required value="<?php echo $rowProduct['quantity']; ?>">
                         </div>
                     </div>
 
                     <div class="form-group format mx-auto inputbox">
-                        <!-- <label for="state" class="col-sm-2 col-form-label mx-auto"></label> -->
                         <select required type="text" class="form-control mx-auto input-type-estoque" name="id_type" value="<?php $rowProduct['type_product_id_type']; ?>">
-                            
+
                             <?php
-                            while ($row = mysqli_fetch_assoc($sqlAllTypes)):
+                            while ($row = mysqli_fetch_assoc($sqlAllTypes)) :
                                 $type = $row['type'];
                                 $id_type = $row['id_type'];
                             ?>
-                                <?php if($rowProduct['type_product_id_type'] == $id_type): ?>  
+                                <!-- Se o tipo selecionado for o tipo do produto mostra o tipo do produto selecionado -->
+                                <?php if ($rowProduct['type_product_id_type'] == $id_type) : ?>
                                     <option selected value="<?php echo $row['id_type']; ?>"><?php echo $row['type'] ?> </option>
-                                
-                                    <?php else:?>
-                                        <option value="<?php echo $row['id_type']; ?>"><?php echo $row['type'] ?> </option>
-                                    <?php endif;?>
 
-                         <?php endwhile;?>
+                                <?php else : ?>
+                                    <!-- Senão mostra todos os tipos para alteração -->
+                                    <option value="<?php echo $row['id_type']; ?>"><?php echo $row['type'] ?> </option>
+                                <?php endif; ?>
+
+                            <?php endwhile; ?>
                         </select>
 
 
@@ -148,11 +151,11 @@
                         </div>
                     </div>
 
-                   
+
                     <a data-bs-toggle="modal" data-bs-target="#addTypeModal"><img src="img/addtype.png" class="add-icon-type-edit-page"> </a>
 
                     <button type="submit" class="btn btn-primary mt-3 ">Confirmar</button>
-                    <a href="estoque2.php"><button type="button" class="btn btn-primary mt-3 ml-3">Cancelar</button></a>
+                    <a href="estoque2.php"><button type="button" class="btn btn-secondary mt-3 ml-3">Cancelar</button></a>
 
                 </form>
 
@@ -162,6 +165,8 @@
             </div>
         </div>
 
+
+
         <div class="modal fade" id="addTypeModal" tabindex="-1" aria-labelledby="addTypeModalLabel" aria-hidden="true">
             <div class="modal-dialog">
                 <div class="modal-content">
@@ -170,8 +175,8 @@
                         <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
                     </div>
                     <div class="modal-body">
-                        <form action="inserttype.php?<?php echo rawurldecode('redirectUrl=edit.php?id='.$id_product)?>" method="POST" enctype="multipart/form-data" class="mx-md-5">
-                        <div class="form-group format mx-auto inputbox">
+                        <form action="inserttype.php?<?php echo rawurldecode('redirectUrl=edit.php?id=' . $id_product) ?>" method="POST" enctype="multipart/form-data" class="mx-md-5">
+                            <div class="form-group format mx-auto inputbox">
                                 <!-- <label for="state" class="col-sm-2 col-form-label mx-auto"></label> -->
                                 <div class="">
                                     <input required type="text" class="form-control mx-auto" name="type" placeholder="Adicionar tipo">
@@ -190,15 +195,20 @@
                                             while ($row = mysqli_fetch_assoc($sqlListTypes)) {
                                                 $id_type = $row['id_type'];
 
-                                                
+
                                                 $i++;
 
                                             ?>
                                                 <tr>
                                                     <th scope="row"><?php echo $i; ?>
-                                                </th>
+                                                    </th>
                                                     <td><?php echo $row['type']; ?></td>
-                                                    <td><a type="button" href="deletetype.php?id=<?php echo $id_type ?>" class="btn btn-danger button-modal-type">Delete</a></td>
+
+                                                    <td><a type="button" data-bs-toggle="modal" data-bs-target="#confirmDeleteModal" href="deletetype.php?id=<?php echo $id_type ?>" class="btn btn-danger button-modal-type">Delete</a></td>
+
+
+
+
                                                 </tr>
 
 
@@ -212,7 +222,6 @@
                                         </tbody>
                                     </table>
                                 </div>
-
                             </div>
 
                             <div class="modal-footer">
@@ -221,6 +230,28 @@
                             </div>
                         </form>
 
+                    </div>
+                </div>
+            </div>
+        </div>
+
+
+        <!-- Modal Confirm Delete Type-->
+        <div class="modal fade" id="confirmDeleteModal" tabindex="-1" aria-labelledby="confirmDeleteModalLabel" aria-hidden="true">
+            <div class="modal-dialog">
+                <div class="modal-content">
+                    <div class="modal-header">
+                        <h1 class="modal-title fs-5" id="confirmDeleteModalLabel">Excluir Tipo</h1>
+                        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                    </div>
+                    <div class="modal-body">
+                    <h3 class="modal-title fs-6">Deseja Excluir Esse Registro? </h3>        
+                          <!-- POP UP SIMPLES <a href="excluir.php" onclick="return confirm('Deseja excluir esse registro ?')">Excluir</a> -->                     
+                    </div>
+
+                    <div class="modal-footer">
+                        <a type="button" href="deletetype.php?id=<?php echo $id_type?>&<?php rawurldecode('redirectUrl=edit.php?id='.$id_product) ?>" class="btn btn-primary button-modal-type">Confirmar</a></td>
+                        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancelar</button>
                     </div>
                 </div>
             </div>
